requirements.txt
tensorflow==2.16.1
pandas 
numpy 
scikit-learn
tensorboard
matplotlib
streamlit
nltk
langchain_community
pypdf
bs4
pymupdf
wikipedia
langchain-text-splitters
langchain-openai
chromadb
sentence_transformers
langchain_huggingface
faiss_cpu
langchain_chroma
langchain_groq
langchain_core
fastapi
langserve
sse_starlette
tf-keras
arxiv
wikipedia
duckduckgo-search
validators==0.28.1
youtube-transcript-api  
unstructured
pytube
numexpr
huggingface_hub
gradio
openai
langchain_nvidia_ai_endpoints
crewai
crewai_tools
pinecone-client
pinecone-text
pinecone-notebooks
neo4j==5.14